remove casting to double

+++ src/feedback/GEAR/feedback.c
@@ -239,7 +239,7 @@ void feedback_will_do_feedback(
     const integertime_t ti_current, const double time_base) {
 
   /* quit if the birth_scale_factor or birth_time is negative */
-  if ((double)sp->birth_scale_factor < 0 || (double)sp->birth_time < 0) 
+  if (sp->birth_scale_factor < 0.0 || sp->birth_time < 0.0) 
     return;
 
   /* quit if the particle contains no SNII */
@@ -308,7 +308,7 @@ void feedback_will_do_feedback(
 int feedback_is_active(const struct spart* sp, const struct engine* e) {
 
   /* the particle is inactive if its birth_scale_factor or birth_time is negative */
-  if ((double)sp->birth_scale_factor < 0 || (double)sp->birth_time < 0) 
+  if (sp->birth_scale_factor < 0.0 || sp->birth_time < 0.0) 
     return 0;
 
   return sp->feedback_data.will_do_feedback;
